requirements.txt
certifi == 2020.6.20
chardet == 3.0.4
config == 0.5.0.post0
idna == 2.10
pip == 20.2.3
python-dateutil == 2.8.1
setuptools == 50.3.0
six == 1.15.0
urllib3 == 1.25.10
requests == 2.24.0
pyTelegramBotAPI == 3.7.3
